修改检视意见

+++ interfaces/innerkits/netfirewallclient/include/netfirewall_common.h
@@ -49,9 +49,9 @@ constexpr int32_t MAX_RULE_PORT_COUNT = 10;
 // Maximum number of domain per rule
 constexpr int32_t MAX_RULE_DOMAIN_COUNT = 100;
 // Maximum exact domain name length
-constexpr int32_t MAX_EXACT_DOMAIN_NAME_LEN = 253;
+constexpr size_t MAX_EXACT_DOMAIN_NAME_LEN = 253;
 // Maximum fuzzy domain name length
-constexpr int32_t MAX_FUZZY_DOMAIN_NAME_LEN = 63;
+constexpr size_t MAX_FUZZY_DOMAIN_NAME_LEN = 63;
 // Intercept log aging: maximum save time
 constexpr int32_t RECORD_MAX_SAVE_TIME = 8 * 24 * 60 * 60;
 // Intercept log aging: Save maximum number of entries
